Unify SuccessResponse and StreamingSuccessResponse
